requirements.txt
crewai 
langchain-community 
pypdf 
faiss-cpu 
rank-bm25 
boto3
python-dotenv
line-bot-sdk
flask